+++ src/torchmetrics/functional/classification/precision_recall_curve.py
@@ -166,6 +166,7 @@ def _binary_precision_recall_curve_format(
     target: Tensor,
     thresholds: Optional[Union[int, list[float], Tensor]] = None,
     ignore_index: Optional[int] = None,
+    normalization: Optional[Literal["sigmoid", "softmax"]] = "sigmoid",
 ) -> tuple[Tensor, Tensor, Optional[Tensor]]:
     """Convert all input to the right format.
 
@@ -182,7 +183,7 @@ def _binary_precision_recall_curve_format(
         preds = preds[idx]
         target = target[idx]
 
-    preds = normalize_logits_if_needed(preds, "sigmoid")
+    preds = normalize_logits_if_needed(preds, normalization)
 
     thresholds = _adjust_threshold_arg(thresholds, preds.device)
     return preds, target, thresholds